chore(ios): update Firebase SDK to 10.4.0
