+++ ios/manifest
@@ -2,7 +2,7 @@
 # this is your module manifest and used by Titanium
 # during compilation, packaging, distribution, etc.
 #
-version: 7.2.0
+version: 8.0.1
 apiversion: 2
 architectures: arm64 x86_64
 description: titanium-firebase-cloud-messaging
