--- ios/platform/FirebaseMessaging.xcframework/ios-arm64_armv7/FirebaseMessaging.framework/Modules/module.modulemap
@@ -1,12 +1,0 @@
-framework module FirebaseMessaging {
-umbrella header "FirebaseMessaging.h"
-export *
-module * { export * }
-  link framework "CoreTelephony"
-  link framework "Foundation"
-  link framework "Security"
-  link framework "SystemConfiguration"
-  link framework "UIKit"
-  link "sqlite3"
-  link "z"
-}
