+++ ios/platform/FirebaseMessaging.xcframework/ios-arm64_x86_64-maccatalyst/FirebaseMessaging.framework/Modules/module.modulemap
@@ -1,5 +1,5 @@
 framework module FirebaseMessaging {
-umbrella header "FirebaseMessaging.h"
+umbrella header "FirebaseMessaging-umbrella.h"
 export *
 module * { export * }
   link framework "CoreTelephony"
