+++ ios/platform/FirebaseMessaging.xcframework/ios-arm64_x86_64-simulator/FirebaseMessaging.framework/Modules/module.modulemap
@@ -1,0 +1,12 @@
+framework module FirebaseMessaging {
+umbrella header "FirebaseMessaging-umbrella.h"
+export *
+module * { export * }
+  link framework "CoreTelephony"
+  link framework "Foundation"
+  link framework "Security"
+  link framework "SystemConfiguration"
+  link framework "UIKit"
+  link "sqlite3"
+  link "z"
+}
