+++ ios/platform/FirebaseMessaging.xcframework/macos-arm64_x86_64/FirebaseMessaging.framework/Modules/module.modulemap
@@ -1,5 +1,5 @@
 framework module FirebaseMessaging {
-umbrella header "FirebaseMessaging.h"
+umbrella header "FirebaseMessaging-umbrella.h"
 export *
 module * { export * }
   link framework "AppKit"
