+++ ios/platform/FirebaseMessaging.xcframework/tvos-arm64/FirebaseMessaging.framework/Info.plist
@@ -13,7 +13,7 @@
 	<key>CFBundlePackageType</key>
 	<string>FMWK</string>
 	<key>CFBundleVersion</key>
-	<string>8.12.0</string>
+	<string>10.4.0</string>
 	<key>DTSDKName</key>
 	<string>iphonesimulator11.2</string>
 </dict>
