+++ ios/platform/FirebaseMessaging.xcframework/tvos-arm64/FirebaseMessaging.framework/Modules/module.modulemap
@@ -1,5 +1,5 @@
 framework module FirebaseMessaging {
-umbrella header "FirebaseMessaging.h"
+umbrella header "FirebaseMessaging-umbrella.h"
 export *
 module * { export * }
   link framework "Foundation"
